+++ src/store.js
@@ -68,6 +68,16 @@ export default class Store {
 	}
 
 	/**
+	 * Returns an array of the names of the keys defined on the object.
+	 *
+	 * @returns {Array.<String>}
+	 * @access public
+	 */
+	keys() {
+		throw new Error('keys() not implemented');
+	}
+
+	/**
 	 * Loads a config file.
 	 *
 	 * @param {String} file - The path to the config file to load.