+++ src/stores/json-store.js
@@ -123,6 +123,16 @@ export default class JSONStore extends Store {
 	}
 
 	/**
+	 * Returns an array of the names of the keys defined on the object.
+	 *
+	 * @returns {Array.<String>}
+	 * @access public
+	 */
+	keys() {
+		return Object.getOwnPropertyNames(this.data);
+	}
+
+	/**
 	 * Loads a config file.
 	 *
 	 * @param {String} file - The path to the config file to load.